Add project reference to JSON3 (JSON polyfill)
SVGMagic depends upon JSON support (it is used internally by jQuery's JSON
deserialisation of the remote server response).  Currently a JSON polyfill
(Douglas Crockford's "json2") is included as part of the SVGMagic script body,
but it could easily be included externally.  This tracks a suitable polyfill
project as a git submodule.