Remove "json2" polyfill from the SVGMagic script body
This was Douglas Crockford's 'json2' polyfill, originally available at:
https://github.com/douglascrockford/JSON-js/blob/master/json2.js

A link to the JSON3 polyfill has already been added to this project.  If JSON is
not supported by the target browser then the JSON polyfill should be included.

This could be accomplished using a framework such as Modernizr, detecting JSON
support and adding the polyfill if it's not there.